USAGOV-2289-chat-paragraph: Add default labels for chat form.
