+++ web/modules/custom/usa_twig_vars/src/Paragraphs/ElectedOfficialsLabels.php
@@ -7,6 +7,8 @@ use Drupal\paragraphs\Entity\Paragraph;
 
 class ElectedOfficialsLabels {
 
+  use mapOverridesTrait;
+
   /**
    * Merges customized labels with defaults for the current languages.
    *
@@ -113,62 +115,6 @@ class ElectedOfficialsLabels {
 
     $defaults = $this->getEmailDefaults($lang);
     return array_replace_recursive($defaults, $overrides);
-  }
-
-  /**
-   * Maps the user submitted values to an array for use by the front-end
-   *
-   * Empty fields are not mapped.
-   *
-   * @param array<string, mixed> $map
-   *
-   * @return array<string, mixed>
-   */
-  private function mapOverrides(Paragraph $para, array $map): array {
-    $overrides = [];
-    foreach ($map as $src => $target) {
-      $value = $para->get($src)->getValue();
-      if (empty($value)) {
-        continue;
-      }
-      $value = trim($value[0]['value']);
-      if (empty($value)) {
-        continue;
-      }
-
-      if (is_array($target)) {
-        // Map user input to a nested array structure.
-        $value = $this->asArray($target, $value);
-        // Need to use array_merge_recursive here to ensure we add sub-keys
-        // to existing values.
-        $overrides = array_merge_recursive($overrides, $value);
-      }
-      else {
-        $overrides[$target] = $value;
-      }
-    }
-
-    return $overrides;
-  }
-
-  /**
-   * Turn an array of keys ands a value into a nested array.
-   *
-   * input  $keys = ['foo', 'bar', 'baz'] and $value = 'Done'
-   * output = ['foo' => ['bar' => ['baz => 'Done']]];
-   *
-   * @param string[] $keys
-   * @param mixed $value
-   *
-   * @return array<string, mixed>
-   */
-  private function asArray(array $keys, $value): array {
-    $key = array_shift($keys);
-    if (count($keys) == 0) {
-      return [$key => $value];
-    }
-
-    return [$key => $this->asArray($keys, $value)];
   }
 
   /**
