+++ web/themes/custom/usagov/templates/paragraphs/paragraph--online-chat-form.html.twig
@@ -1,15 +1,13 @@
 {% set chatID = elements.field_chat_id|field_value %}
-
 <div id="liveagent_button_online_{{ chatID }}">
   {{ elements.field_chat_online }}
   <button id="chat-button"
           href="javascript://Chat"
           onclick="liveagent.startChat('{{ chatID }}')"
           class="usa-button"
-    >{{ elements.field_chat_button|field_value }}
+    >{{ chatLabels['chat-button'] }}
   </button><br />
-  <br />
-  &nbsp;
+  <br />&nbsp;
   {{ elements.field_chat_online_footer }}
 </div>
